Gravity Field
I've worked on the Zero gravity area in blueprint, and have completed it to a basic standard, It does what it needs to do however right now its hard coded to work with one specific actor and that will be the issue I will overcome next.

The current way to toggle gravity is by entering the area, this can be later changed to being done on creation when the berries are hit.

I've also made it into a box collision for now, this will be a quick change into sphere.