Added in Teleport plant blueprint, needs further work
How to use:
The teleport plant blueprint is now in the project and can be used by placing two of them in the scene and linking them up via the "Linked Plant" variable within the actors details. The Teleport point on the plant can also be altered to change the position and rotation the character will appear at.

Other notable variables that can be changed are:
Teleport delay - time before teleporting the player
Cooldown Duration - the time after teleporting before the teleporter can be used again. (needed to prevent endless teleporting and allows the player to stand on the teleported plant for as long as they want and needing them to move off and back on to teleport again).

TODO:
The plant still needs additional checks to prevent errors and the addition of stopping player movement when entering it.