Gravity Field Update
I have updated my anti-gravity area to now work based fully off the tags feature, The actor must have a type of collision box/sphere/capsule (Should be a given since gravity needs collision) or it will return an error. It collects all the tagged actors in the collision and then loops through checking if it has any of the collisions before turning off their gravity using that information, after X amount of time it will return from floating and removes the list of affected actors to then be able to be set again. - Also there is a lock stopping the code from running multiple times until it is finished.

I have also made a few public variables to make it easier to change stuff around, these include the impulse force for when the objects have gravity removed to simulate floating away, the amount of time they stay in anti-gravity and the tag required for them to be affected.

I changed some of the blueprints around by adding an event to switch gravity using some inputs so that there isn't too much blueprint.