created interaction mechanic
A test object is included with an example of how the interface can be implemented in C++. for some reason if the interface is called on the player or an object that doesn't implement it it will crash the engine. This shouldn't be an issue provided all interactions are done through the provided input in the character. but if not make sure to check for both those things.