Deer Implemented
f the player approaches it, it runs away, but if the player stands still for 5 seconds, the deer will approach the player.

The Deer has a lot of options to change its values in the behaviour tree, I may change some of them to blackboard variable so they are easily changeable and therefore making it easier for designers to tweak.