Pause menu WIP
There is a pause menu in place that can be activated by pressing P, currently the resume function works as well as pressing exit, this will cause a confirm menu to appear where only the no button works.

To do this I've added a new input action for the P key and added some blueprint in the player character for it to be activated. Along with two UI widgets for the menus.