The quit confirm functional + restart WIP
Added functionality to the quit confirmation button. Pressing will now just end the game.

The restart part of the pause menu will also now open up a separate menu for confirmation. The no option works and resumes the game, the yes option however is still WIP and will reset the level but the player looses controls.