Inventory system partially working
There is now a toggleable inventory panel which can be viewed with the "I" key. Once you pickup an inventory Item, the data from the item itself will include a texture that will be used to represent the item in the inventory.  The interactable interface is executed on each of the three types of items. They always execute this as they always need to "do" something, either destroy itself or perform some functionality. Multiple Items can also be picked up / interacted with at once.