Bug Creature
Today I have worked on the Bug, the waypoints for the bug and the nest. The way the bug works is simply you interact with it, and then it goes to the nest, I have implemented waypoints that can be placed into the scene so that the bug can follow a set path and look more like a bug. I also added a reset feature in case the bug should respawn where it started with a custom amount that if set to Zero will not reset the bug.