Basic Rock/Snail creature made I always get confused with the name along side a basic implementation of a object spawner volume
Currently with the object spawner you can set the amount of times the test object is spawned. this object can be moved towards the snail and the snail would then output a debug string in satisfaction. the object which is spawned can be changed as well. I plan on pooling the objects together so I can also keep track of the amount of iterations within the scene so I can automate spawning of the objects by how many still exist within the object pool.