Berry Bird Hop Finished
Added a Hop after the bird is on the ground. I had some issues with AI movement but figured out the issue lied in simulate physics being on for the launch out of the bush stopping AiMove from working, setting this back to false when resetting the bird to upright fixed the issue.